requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
packaging==25.0
pdf2image==1.17.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytesseract==0.3.13
python-decouple==3.8
sqlparse==0.5.3
